second v0.1.0 Identicon

Functions

buildPixelGrid(structure)

build_grid(structure)

this method is the abbrevation of

def build_grid(structure) do
    %Identicon.Image{input_in_binary: binary} = structure;
    arrayOfArray = Enum.chunk(binary, size_chunk);
    arrayOfArray;

This -> Enum.map(fn(row)-> mirrorGrid(row) end) it’s the same that the following

    arrayOfArrays = binary |> Enum.chunk(3)
    array = [] 
    for row <- arrayOfArrays do
        array ++ mirrorGrid(row)
    end

A better practice to Enum.map(fn(row)-> mirrorGrid(row) end) it’s:

Enum.map(&mirrorGrid/1);

    & <- specific that a value will be pass to mirrorGrid
    mirrorGrid/1 <- specific that the method to run is called mirrorGrid and just have an argument

convertToHash(input)

This method will get a string and later encrypt this to then pass it to binary
finally this will return a struct with the valor in binary

drawImage(image)

getEvenSquares(structure)

main(input)

mirrorGrid(row)

newTermsLearned(input \\ "Jesus")

pickColor(binary_value)

This will have a binary_value, to then do match and assign this to a variable called binary_list
finally we will get the first three value using pipe operator 
"Give the first three values and after pipe i know that exist more values, but ignore this"
Remember that underscore followed by a variable, will be ignored

This method will return something as this

    %Identicon.Image{color_rgb: {233, 130, 150},
    input_in_binary: [233, 130, 150, 8, 221, 144, 255, 107, 139, 247, 203, 80, 116,110, 174, 120]}

This is the same as the method

    def pickColor(binary_value) do
        %Identicon.Image{hex: binary_list} = binary_value
        [r, g, b | _rest] = binary_list
        %Identicon.Image{binary_value | color: {r,g,b}} 
        # <- the binary_value do as the spread operator ...binary_value
        [r,g,b] #<- this will return the first three values
    end 

saveImage(image, filename)